backend/export-code-word/requirements.txt
python-docx>=1.1.0
boto3>=1.26.0
